# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events/{slug}": {
      "get": {
        "operationId": "EventsController_bySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar evento pelo slug",
        "tags": [
          "Events"
        ]
      }
    },
    "/api/v1/categories": {
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar categorias ativas",
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Criar categoria de evento",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualizar categoria de evento",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/organizer/events": {
      "get": {
        "operationId": "OrganizerEventsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar eventos do organizador",
        "tags": [
          "Organizer"
        ]
      },
      "post": {
        "operationId": "OrganizerEventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Criar evento em rascunho",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}": {
      "patch": {
        "operationId": "OrganizerEventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualizar evento do organizador",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}/publish": {
      "post": {
        "operationId": "OrganizerEventsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Publicar evento",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}/cancel": {
      "post": {
        "operationId": "OrganizerEventsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Cancelar evento e encerrar vendas",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}/ticket-types": {
      "post": {
        "operationId": "OrganizerEventsController_createTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Criar tipo e unidades de ingresso",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{eventId}/ticket-types/{ticketTypeId}": {
      "patch": {
        "operationId": "OrganizerEventsController_updateTicket",
        "parameters": [
          {
            "name": "ticketTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualizar tipo de ingresso",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{eventId}/ticket-types/{ticketTypeId}/capacity": {
      "patch": {
        "operationId": "OrganizerEventsController_capacity",
        "parameters": [
          {
            "name": "ticketTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCapacityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Alterar capacidade do tipo de ingresso",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}/images/cover": {
      "post": {
        "operationId": "OrganizerEventsController_cover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Enviar capa do evento",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{id}/images/gallery": {
      "post": {
        "operationId": "OrganizerEventsController_gallery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Adicionar imagem à galeria",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{eventId}/images/order": {
      "patch": {
        "operationId": "OrganizerEventsController_reorderImages",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderImagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reordenar imagens da galeria",
        "tags": [
          "Organizer"
        ]
      }
    },
    "/api/v1/organizer/events/{eventId}/images/{imageId}": {
      "delete": {
        "operationId": "OrganizerEventsController_removeImage",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remover imagem do evento",
        "tags": [
          "Organizer"
        ]
      }
    },
...
    "/api/v1/admin/organizer-invitations": {
      "post": {
        "operationId": "InvitationsController_inviteOrganizer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Convidar um organizador",
        "tags": [
          "Invitations"
        ]
      },
      "get": {
        "operationId": "InvitationsController_listOrganizer",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar convites de organizador",
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/invitations/organizer/accept": {
      "post": {
        "operationId": "InvitationsController_acceptOrganizer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Aceitar convite de organizador",
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/organizer/events/{eventId}/staff-invitations": {
      "post": {
        "operationId": "InvitationsController_inviteStaff",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Convidar membro da portaria",
        "tags": [
          "Invitations"
        ]
      },
      "get": {
        "operationId": "InvitationsController_listStaff",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Listar convites de portaria do evento",
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/invitations/staff/accept": {
      "post": {
        "operationId": "InvitationsController_acceptStaff",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Aceitar convite de portaria",
        "tags": [
          "Invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ingressos Locais API",
    "description": "API transacional de eventos, reservas, pedidos e validação de ingressos.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      },
      "idempotency-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 180
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 220,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 10000
          },
          "venueName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 180
          },
          "postalCode": {
            "type": "string",
            "pattern": "^\\d{8}$"
          },
          "street": {
            "type": "string",
            "minLength": 2,
            "maxLength": 180
          },
          "number": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          },
          "complement": {
            "type": "string",
            "maxLength": 100
          },
          "district": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "city": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "state": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "timezone": {
            "default": "America/Sao_Paulo",
            "type": "string"
          }
        },
        "required": [
          "categoryId",
          "title",
          "slug",
          "description",
          "venueName",
          "postalCode",
          "street",
          "number",
          "district",
          "city",
          "state",
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 180
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 10000
          },
          "venueName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 180
          },
          "postalCode": {
            "type": "string",
            "pattern": "^\\d{8}$"
          },
          "street": {
            "type": "string",
            "minLength": 2,
            "maxLength": 180
          },
          "number": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          },
          "complement": {
            "type": "string",
            "maxLength": 100
          },
          "district": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "city": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "state": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "timezone": {
            "default": "America/Sao_Paulo",
            "type": "string"
          }
        }
      },
      "CreateTicketTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "priceCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000000
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "maxPerOrder": {
            "default": 10,
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "saleStartsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "saleEndsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "name",
          "priceCents",
          "capacity"
        ]
      },
      "UpdateTicketTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "priceCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000000
          },
          "maxPerOrder": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "saleStartsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "nullable": true
          },
          "saleEndsAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "UpdateCapacityDto": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000
          }
        },
        "required": [
          "capacity"
        ]
      },
      "ReorderImagesDto": {
        "type": "object",
        "properties": {
          "imageIds": {
            "minItems": 1,
            "maxItems": 6,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "imageIds"
        ]
      },
...
      "InviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 32,
            "maxLength": 500
          }
        },
        "required": [
          "token"
        ]
      }
    }
  }
}